Extract workspace from main
Handles window-specific configuration
